#25535 moved command logic to separate modules using commandDir
